자산추가 중복일때 총자산 표시, 총자산 Int로 수정

+++ TradingRecordApp/Asset/AssetDetailViewController.swift
@@ -96,7 +96,7 @@ extension AssetDetailViewController: UITableViewDelegate, UITableViewDataSource 
         guard let cell = self.tableView.dequeueReusableCell(withIdentifier: "AssetDetailListCell") as? AssetDetailListCell else { return UITableViewCell() }
         
         let coinPriceDouble = Double(self.assetDetailList[indexPath.row].coinInfo.coinPrice)
-        let evalPrice = self.assetDetailList[indexPath.row].coinAmount * coinPriceDouble!
+        let evalPrice = Int(self.assetDetailList[indexPath.row].coinAmount * coinPriceDouble!)
         
         cell.coinNameDetail.text = self.assetDetailList[indexPath.row].coinName
